genera las local notifications y redirige segun su accion
